fix test module imports through symlinks, write first test for semantic checker
Will need to start merging because extensions not supported by Cuckoo
